+++ cli/custom/groups.go
@@ -36,7 +36,6 @@ var commandGroup = map[string]string{
 	"auth":       groupGetStarted,
 	"login":      groupGetStarted,
 	"logout":     groupGetStarted,
-	"whoami":     groupGetStarted,
 	"status":     groupGetStarted,
 	"switch":     groupGetStarted,